Extend ground class to a building object
