+++ src/main/java/com/prj/commom/factory/BoardSeeder.java
@@ -5,15 +5,14 @@ import com.org.chained_list.DoubleChainedList;
 import com.prj.commom.BoardNode;
 import com.prj.commom.Dice;
 import com.prj.entity.PathBoard;
-import com.prj.entity.Ground;
+import com.prj.entity.building.Ground;
 import com.prj.entity.Player;
-
-import java.util.ArrayList;
 
 public class BoardSeeder {
     public static int SIZE = 38;
     public static int TOTAL_PROPORTION = 10;
     public static int GROUND_PROPORTION = 7;
+    public static int GROUND_AMOUNT_PERCENT = 2;
     public static long GROUND_PRICE_MULTIPLIER = 100;
     public static int GROUND_PRICE_RANGE = 90;
 
@@ -45,8 +44,9 @@ public class BoardSeeder {
         int randBoundNumber = Dice.nextInt(prices.getSize());
         int randPrice = prices.get(randBoundNumber).getValue();
         long price = Dice.nextPositiveInt(randPrice) * GROUND_PRICE_MULTIPLIER;
+        int rentAmount = (GROUND_AMOUNT_PERCENT * randPrice) / 10;
 
-        Ground ground = new Ground(price, banker);
+        Ground ground = new Ground(price, rentAmount, banker);
         return new BoardNode(ground);
     }
 }
