+++ src/test/java/com/prj/entity/bank/BankerTest.java
@@ -2,12 +2,10 @@ package com.prj.entity.bank;
 
 import com.org.chained_list.DoubleChainedList;
 import com.prj.TestCase;
-import com.prj.entity.Ground;
+import com.prj.entity.building.Ground;
 import com.prj.entity.Player;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-
-import java.util.ArrayList;
 
 import static org.junit.jupiter.api.Assertions.*;
